+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Change Log
 
+## [0.3.0-beta] - Feb 27, 2025
+- Exposed SimpleTimer functionality for Rewind and FastForward in the Timer class
+
 ## [0.2.1-beta] - Feb 26, 2025
-- reduced the ServiceLocator version support to 0.5.0
+- Reduced the ServiceLocator version support to 0.5.0
 
 ## [0.2.0-beta] - Feb 26, 2025
 - Added extension support for ServiceLocator